fix: dashboard copy bug fix

+++ server/src/main/java/datart/server/base/dto/DashboardBaseInfo.java
@@ -17,7 +17,7 @@ public class DashboardBaseInfo {
 
     private Boolean isFolder;
 
-    private Integer index;
+    private Double index;
 
     public DashboardBaseInfo(Dashboard dashboard) {
         BeanUtils.copyProperties(dashboard, this);
